requirements.txt
Pillow==8.1.0
Django==3.1.6
protobuf==3.14.0
numpy==1.20.1
object_detection==0.0.3
pytesseract==0.3.7
djangorestframework==3.12.2
six==1.15.0
tensorflow==2.4.1
